// server/src/enum/http-code.enums.ts
export enum HttpCode {
  OK = 200,
  Created = 201,
  Accepted = 202,
  Non_Authoritative_Information = 203,
  No_Content = 204,
  Reset_Content = 205,
  Not_Modified = 304,
  BadRequest = 400,
  Unauthorized = 401,
  Payment_Required = 402,
  Forbidden = 403,
  Not_Found = 404,
  Method_not_Allowed = 405,
  Not_Acceptable = 406,
  Conflict = 409,
  Internal_server_Error = 500,
  Bad_Gateway = 502,
}